Hm, layout problem with 64-bit systems now we've changed to compiling with XCode 4.
I think this might be related to how the new NSUInteger types are used; however, the warnings that are coming back are really not very helpful.

+++ GlkView/GlkTextWindow.m
@@ -309,7 +309,7 @@
 
 - (BOOL) textView: (NSTextView*) view
 	clickedOnLink: (id) link 
-		  atIndex: (unsigned) charIndex {
+		  atIndex: (NSUInteger) charIndex {
 	if ([link isKindOfClass: [NSNumber class]]) {
 		if (hyperlinkInput) {
 			// Generate the event for this hyperlink
